make redirect from evn

+++ resources/views/group/index.blade.php
@@ -13,6 +13,7 @@
                         <tr>
                             <th>Title</th>
                             <th>Description</th>
+                            <th>Owner</th>
                             <th>Action</th>
                         </tr>
                         </thead>
@@ -22,9 +23,18 @@
                                 <td><a href="{{ route('password.index',['group'=>$group]) }}">{{ $group->name }}</a></td>
                                 <td>{{ $group->description }}</td>
                                 <td>
+                                    @foreach($group->owners as $owner)
+                                        {{ $owner->name }}<br>
+                                    @endforeach
+                                </td>
+                                <td>
                                     <a href="{{ route('password.index',['group'=>$group]) }}" class="btn btn-sm btn-info" title="View"><i class="fa fa-eye"></i></a>
-                                    <a href="{{ route('group.edit', ['id'=>$group->id]) }}" class="btn btn-sm btn-warning" title="Edit"><i class="fa fa-pencil"></i></a>
-                                    <a href="{{ route('group.destroy', ["id" => $group->id]) }}" class="btn btn-sm btn-danger"  data-method="DELETE" data-confirm="Are you sure ?" data-token="{{ csrf_token() }}"><i class="fa fa-trash"></i></a>
+                                    @can("update", $group)
+                                        <a href="{{ route('group.edit', ['id'=>$group->id]) }}" class="btn btn-sm btn-warning" title="Edit"><i class="fa fa-pencil"></i></a>
+                                    @endcan
+                                    @can("destroy", $group)
+                                        <a href="{{ route('group.destroy', ["id" => $group->id]) }}" class="btn btn-sm btn-danger"  data-method="DELETE" data-confirm="Are you sure ?" data-token="{{ csrf_token() }}"><i class="fa fa-trash"></i></a>
+                                    @endcan
                                 </td>
                             </tr>
                         @endforeach
